This reverts PR #13 and puts us back on SGX 2.16
This is a bit funky but doing this allows us to get test coverage
on the go-grpc-gateway in master, without also bumping SGX to 2.17.

We can revert this and so increase to SGX 2.17 change at a later
time when we are ready to do that.

In the meantime, it's bad for the docker image to keep progressing
in versions without any of them being tested in CI.